+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+**20160718**
+
+- Created mechanism for examples with test data indirection
+- Scripts to initialize PowerShellGet
+- Scripts to initialize remoting with CredSSP
+
 **20160712**
 
 - Initial commit. Describing the repository